libs/agent-core/pyproject.toml
[project]
name = "agent-core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Julian Hecker" }]
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.17.0",
    "pydantic>=2.12.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
